libcrux: Move hash into separate module
